refactor: RunningService -> RecordService 변경
내부 작업은 기록과 동일하기에 파일 변경
