+++ src/main/kotlin/com/yapp/yapp/record/domain/record/RunningRecordManager.kt
@@ -2,8 +2,8 @@ package com.yapp.yapp.record.domain.record
 
 import com.yapp.yapp.common.exception.CustomException
 import com.yapp.yapp.common.exception.ErrorCode
+import com.yapp.yapp.record.domain.Pace.Companion.averagePace
 import com.yapp.yapp.record.domain.point.RunningPoint
-import com.yapp.yapp.running.domain.Pace.Companion.averagePace
 import com.yapp.yapp.running.domain.RunningMetricsCalculator.roundTo
 import org.springframework.stereotype.Component
 import java.time.OffsetDateTime
@@ -37,7 +37,7 @@ class RunningRecordManager(
         userId: Long,
     ): RunningRecord {
         val runningRecord = getRunningRecord(id, userId)
-        runningRecord.finish()
+        runningRecord.stop()
         return runningRecord
     }
